edited new file

+++ src/transactions/store/TransactionServices.js
@@ -1,15 +1,9 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
+import { GetSettings } from '../../settings/store/SettingsServices';
 
-const { ipcRenderer } = window.require('electron');
-let TransServices;
-
-ipcRenderer.on('get-ip',(e,args)=>{
-    const host = args.address ? args.address : 'localhost';
-    TransServices = axios.create({
-        baseURL : `http://${host}:8081/api/v1`,
-        timeout : 60 * 2 * 1000
-    });
+const TransServices = axios.create({
+    timeout : 60 * 2 * 1000
 });
 
 const sleep = (x)=>{
@@ -18,15 +12,25 @@ const sleep = (x)=>{
 
 export const getAllTransaction = createAsyncThunk(
     'transactions/getAllTransaction',
-    async(args,{ rejectWithValue })=>{
+    async(args,{ rejectWithValue,dispatch })=>{
         const { opt } = args;
         try{
-            const res = await TransServices({
-                ...opt,
-                method : 'GET'
-            });
-            await sleep(2000);
-            return res.data;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'GET'
+                });
+                await sleep(2000);
+                return res.data;            
+            }            
         }catch(err){
             return rejectWithValue(err);
         }
@@ -35,15 +39,25 @@ export const getAllTransaction = createAsyncThunk(
 
 export const getSingleTransaction = createAsyncThunk(
     'transactions/getSingleTransaction',
-    async(args, {rejectWithValue})=>{
+    async(args, {rejectWithValue,dispatch})=>{
         const { opt } = args;
         try{
-            const res = await TransServices({
-                ...opt,
-                method : 'GET'
-            });
-            await sleep(2000);
-            return res.data;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'GET'
+                });
+                await sleep(2000);
+                return res.data;            
+            }            
         }catch(err){
             return rejectWithValue(err);
         }
@@ -52,16 +66,26 @@ export const getSingleTransaction = createAsyncThunk(
 
 export const createTransaction = createAsyncThunk(
     'transactions/createTransaction',
-    async(args,{ rejectWithValue })=>{
+    async(args,{ rejectWithValue,dispatch })=>{
         const { opt,values } = args;
         try{
-            const res = await TransServices({
-                ...opt,
-                method : 'POST',
-                data : values
-            });
-            await sleep(2000);
-            return res.data;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'POST',
+                    data : values
+                });
+                await sleep(2000);
+                return res.data;            
+            }            
         }catch(err){
             return rejectWithValue(err);
         }
@@ -70,15 +94,25 @@ export const createTransaction = createAsyncThunk(
 
 export const findTransaction = createAsyncThunk(
     'transactions/findTransaction',
-    async( args, { rejectWithValue } )=>{        
+    async( args, { rejectWithValue,dispatch } )=>{        
         try{
-            const { opt } = args;
-            const res = await TransServices({
-                ...opt,
-                method : 'GET'
-            });
-            await sleep(2000);
-            return res.data;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const { opt } = args;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'GET'
+                });
+                await sleep(2000);
+                return res.data;            
+            }            
         }catch(err){
             return rejectWithValue(err.response.data);
         }
@@ -87,21 +121,31 @@ export const findTransaction = createAsyncThunk(
 
 export const getDocDef = createAsyncThunk(
     'transactions/getDocDef',
-    async(args,{rejectWithValue})=>{
+    async(args,{rejectWithValue,dispatch})=>{
         try{
-            const { opt } = args;
-            const res = await TransServices({
-                ...opt,
-                method : 'GET',
-                responseType : 'blob',  
-                headers: {
-                    Accept: 'application/pdf',
-                    'Content-Type': 'application/pdf',
-                    mode : 'no-cors'
-                }
-            });
-            const url = window.URL.createObjectURL(res.data);
-            return url;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const { opt } = args;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'GET',
+                    responseType : 'blob',  
+                    headers: {
+                        Accept: 'application/pdf',
+                        'Content-Type': 'application/pdf',
+                        mode : 'no-cors'
+                    }
+                });
+                const url = window.URL.createObjectURL(res.data);
+                return url;
+            }            
         }catch(err){
             return rejectWithValue(err.response.data);
         }
@@ -110,16 +154,26 @@ export const getDocDef = createAsyncThunk(
 
 export const updateTransaction = createAsyncThunk(
     'transactions/updateTransaction',
-    async(args,{rejectWithValue})=>{
+    async(args,{rejectWithValue,dispatch})=>{
         try{
-            const { opt,value } = args;
-            const res = await TransServices({
-                ...opt,
-                method : 'PATCH',
-                data : value
-            });
-            await sleep(2000);
-            return res.data;
+
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const { opt,value } = args;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'PATCH',
+                    data : value
+                });
+                await sleep(2000);
+                return res.data;            
+            }            
         }catch(err){
             return rejectWithValue(err.response.data);
         }
@@ -128,15 +182,25 @@ export const updateTransaction = createAsyncThunk(
 
 export const deleteTransaction = createAsyncThunk(
     'transactions/deleteTransaction',
-    async(args,{rejectWithValue})=>{
+    async(args,{rejectWithValue,dispatch})=>{
         try{
-            const { opt } = args;
-            const res = await TransServices({
-                ...opt,
-                method : 'DELETE'
-            });
-            await sleep(2000);
-            return res.data;
+            
+            const resSettings = await dispatch( GetSettings({
+                url : "/settings"
+            }) );
+
+            if(GetSettings.fulfilled.match(resSettings)){
+                const { settings } = resSettings.payload;
+                const host = settings.address !== undefined ? settings.address : undefined;
+                const { opt } = args;
+                const res = await TransServices({
+                    ...opt,
+                    baseURL : `http://${host}:8081/api/v1`,
+                    method : 'DELETE'
+                });
+                await sleep(2000);
+                return res.data;
+            }
         }catch(err){
             return rejectWithValue(err.response.data);
         }
